feat(file-args): adding the ability to specify a file arg
When objects are much more complex it becomes tedius to type out the
entire json string to represent an object. Adding the ability to add an
argument of the format: f:"Product.json" to read a json file and pass
that as the argument to the contract client.

Breaks nothing