feat(custom-args): adding the ability to use js dicts
We previously passed custom types as a string that needed to be parsed
for custom types. This was easier with the CLI but is somewhat clunky in
the browser. We now accept both string and javascript dicts as
arguments. To make this possible I changed the 'file' argument used by
the CLI to be stored at the _type keyword if being used.